+++ antifraud/src/main/java/com/bank/antifraud/entity/AuditEntity.java
@@ -4,7 +4,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 /**
@@ -22,7 +28,7 @@ public class AuditEntity {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    BigInteger id;
     @Column(name = "entity_type",
             nullable = false)
     String entityType;
